Fix a bug in protorpc. This is possibly a breaking change.
Previously, when creating a TimeZoneOffset from a datetime.timedelta, protorpc
would use the total number of seconds for computing the offset. However, all
the docs and code consuming this offset were expecting it to be calculated
in *minutes*, not seconds.

I don't know if this will wreak havoc on someone in the wild who was already
working around this.