Switch `Field.__metaclass__` to `six.with_metaclass`.
There was still one `__metaclass__` hiding, which was of course ignored in
python3. Oops.

I think I'm actually going to do a version number bump after this, since we're
breaking a public interface. In principle, we could keep `Field.__metaclass__`
for people to use, but it's going to break in python3 anyway -- no reason to
mislead.